// src/interface/c-financing-purchase-section-result.interface.ts
export interface CFinancingPurchaseSectionResultInterface {
    label: string;
    purchasePrice: number;
    downPayment: number;
    interestRate: number;
    loanLengthYears: number;
    closingCosts: number;
    rehabCosts: number;
    monthlyIncome: number;
    monthlyFixedExpenses: number;
    loanAmount: number;
    mortgagePaymentPI: number;
    cashToClose: number;
    totalMonthlyCashFlow: number;
    yearlyCashFlow: number;
}