// src/interface/c-financing-five-year-snapshot-result.interface.ts
export interface CFinancingFiveYearSnapshotResultInterface {
    cumulativeNetCashFlow: number;
    totalPrincipalReduction: number;
    appreciationGain: number;
    estTaxBenefits: number;
    totalCumulativeWealth: number;
    annualizedROI: number;
    cumulativeWealthWithInitialInvestment: number;
}